refactored bullet into its own module
