+++ grillBullet/src/main/java/module-info.java
@@ -1,0 +1,7 @@
+module grillBullet {
+    exports grillBullet;
+    requires common;
+    requires java.desktop;
+
+    provides interfaces.IPlugin with grillBullet.Bullet;
+}
