+++ pom.xml
@@ -25,6 +25,7 @@
         <module>player</module>
         <module>movement</module>
         <module>grill</module>
+        <module>grillBullet</module>
     </modules>
 
     <dependencies>
